org: move PipelineCalcuator into its own file, separate from page.tsx due to complexity
